feat(core): allow scenarios to start with When or Then directly
- Add when() and then() methods to BddScenario class
- Extract BddScenario to separate bdd_scenario.dart part file
- Add comprehensive doc comments with examples
- Add bdd_scenario_test.dart for verification

Closes #14
